Run unit specks whenever we are going to run unit tests.
This is a strategy change. It makes more to transparently run the specks if running unit tests because most users are unaware of the “phases” of the test process.

To be clear, this means that `grails test-app --unit` will now invoke the specks as well,  where as previously it would not.